+++ packages/widgetbook_golden_test_core/CHANGELOG.md
@@ -4,6 +4,8 @@
 - Added [precacheImagesTimeout] property to [WidgetbookGoldenTestsProperties] to configure the timeout for precaching images.
 - Added [tags] property to [WidgetbookGoldenTestsProperties] to configure the tags for the golden tests. Also added a [tags] property to [WidgetbookGoldenTestBuilder] to configure the tags per use-case.
 - [GoldenPlayAction] now has a [skip] property to configure the skip per golden play action instead of relying on the [skip] property of [WidgetbookGoldenTestBuilder].
+- [GoldenPlayAction] now has a [customPump] property to execute a custom pump function after the action instead of hard-coded `pumpAndSettle()`.
+- [WidgetbookGoldenTestBuilder] now has [pumpBeforeImagePrecache] and [pumpAfterImagePrecache] properties to configure custom pump functions before and after image precaching.
 
 ### Fixed
 - Use runZonedGuarded instead of simple try catch when trying to extract [WidgetbookGoldenTestBuilder] metadata. This solves an issue where no test were being executed when there was an error during the metadata extraction.